requirements.txt
langchain==0.1.13
langchain-community==0.0.29
langchain-core==0.1.33
langchain-openai==0.1.1
langchain-text-splitters==0.0.1
langsmith==0.1.31
nltk==3.8.1
openai==1.14.2
poppler-utils==0.1.0
pymilvus==2.4.0
pypdf==4.1.0
python-dotenv==1.0.1
tiktoken==0.6.0